Deduplicate approve_and_execute in tests
